Migrate colour foundations
